+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+<a name="1.2.0"></a>
+# [1.2.0](https://github.com/raulsebastianmihaila/create-crizmas/compare/v1.1.0...v1.2.0) (2020-11-30)
+
+### Updates
+- Add loader for css to ensure minification and cache busting.
+- Add html lang attribute.
+- Update the version of crizmas-components.
+
 <a name="1.1.0"></a>
 # [1.1.0](https://github.com/raulsebastianmihaila/create-crizmas/compare/v1.0.0...v1.1.0) (2020-09-29)